setup for ascend dev
