+++ ascend_plan.md
@@ -1,0 +1,33 @@
+# Adapting Codegen from CUDA to AscendC for Huawei Hardware
+
+AscendC is Huawei's C++-based domain-specific language for programming Ascend AI processors (NPUs), analogous to CUDA for NVIDIA GPUs. It supports kernels, memory hierarchies (e.g., unified buffer, local memory), and parallelism via cores/clusters, but with differences in syntax, memory models, and execution hierarchy (e.g., no direct equivalent to CUDA's grid/block/thread; instead, it uses task-based scheduling and vectorized operations). Adapting Descend's codegen to target AscendC is feasible since the frontend (parsing and type checking) is backend-agnostic, but it would require significant modifications to the `codegen` module. Here's how you could go about it:
+
+1. **Fork and Understand the Repo**:
+   - Clone <https://github.com/trusted-programming/descend>.
+   - Study the `codegen` module (in `src/codegen/`) to identify the CUDA-specific parts: the CUDA AST definitions, translation logic (likely in a visitor or recursive function traversing the Descend AST), and printer.
+
+2. **Create a New Backend**:
+   - Add a new submodule, e.g., `src/codegen/ascend/`, or refactor `codegen` to support multiple targets via a flag (e.g., `--target=ascend` in the compiler CLI).
+   - Define an AscendC-specific AST (e.g., in `ascend_ast.rs`) with nodes for AscendC elements: kernels (using `__global__` or task functions), memory ops (e.g., `aclrtMalloc` equivalents), indices (core IDs, vector lanes), and expressions.
+
+3. **Modify Translation Logic**:
+   - Reuse the inlining step as-is, since it's general.
+   - Map Descend constructs to AscendC:
+     - **Execution Contexts**: `gpu.grid`/`gpu.block`/`gpu.thread` to Ascend's core/cluster/task model. For example, translate `sched` to AscendC's parallel loops or vector intrinsics (e.g., using `davinci::for_each` or similar for data parallelism).
+     - **Memory Management**: `gpu.global` to Ascend's global/unified memory (`__global__` pointers); `shrd`/`uniq` borrows to ensure safe access, but drop annotations as in CUDA. Built-ins like `gpu_alloc_copy` to Ascend APIs (e.g., `aclrtMemcpy` for copies).
+     - **Views and Indexing**: Similar reverse-order index combination, but using AscendC's tensor/view APIs if available, or manual index calculations.
+     - **Kernels and Launches**: `exec` to AscendC kernel enqueue (e.g., via `aclopExecute` or task launches). Adjust for Ascend's lack of explicit grid/block dims—use its operator fusion or graph-based execution.
+     - Handle differences: AscendC emphasizes AI ops (e.g., matrix multiplies), so optimize for that; add support for Ascend's vector types (e.g., `float16x8`).
+   - Implement this in a translator function (e.g., recursive over the Descend AST nodes), similar to CUDA's but with AscendC mappings.
+
+4. **Update Emission**:
+   - Write a printer for the AscendC AST to generate AscendC C++ code strings, including Huawei-specific headers (e.g., `davinci_api.h`).
+   - Ensure output compiles with the Ascend compiler (`asc` tool).
+
+5. **Testing and Integration**:
+   - Add AscendC examples in a new directory (e.g., `ascend-examples/`).
+   - Test on Huawei hardware; adjust for performance (AscendC often requires manual optimization for pipelines/vectorization).
+   - Potential challenges: AscendC's closed ecosystem (need Huawei SDK), differences in error handling, and less mature tooling compared to CUDA.
+
+This would involve ~1000-2000 lines of Rust changes, depending on complexity. If you're familiar with AscendC, start by prototyping a simple example (e.g., vector scale) manually in AscendC, then reverse-engineer the mappings. For community help, open issues on the repo or discuss in related forums.
+<argument name="citation_id">3</argument>
