+++ src/codegen/ascend/mod.rs
@@ -1,0 +1,5 @@
+use crate::ast as desc;
+
+pub fn gen(comp_unit: &desc::CompilUnit, idx_checks: bool) -> String {
+    unimplemented!()
+}
