+++ src/lib.rs
@@ -8,9 +8,17 @@ pub mod error;
 pub mod parser;
 pub mod ty_check;
 
-pub fn compile(file_path: &str) -> Result<String, ErrorReported> {
+pub enum Backend {
+    Cuda,
+    AscendC,
+}
+
+pub fn compile(file_path: &str, backend: Backend) -> Result<String, ErrorReported> {
     let source = parser::SourceCode::from_file(file_path)?;
     let mut compil_unit = parser::parse(&source)?;
     ty_check::ty_check(&mut compil_unit)?;
-    Ok(codegen::gen(&compil_unit, false))
+    match backend {
+        Backend::Cuda => Ok(codegen::cuda::gen(&compil_unit, false)),
+        Backend::AscendC => Ok(codegen::ascend::gen(&compil_unit, false)),
+    }
 }
