+++ tests/example_infer_apps.rs
@@ -8,7 +8,7 @@ type Res = Result<(), descend::error::ErrorReported>;
 fn transpose() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/transpose.desc")?
+        descend::compile("examples/infer/transpose.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -16,7 +16,10 @@ fn transpose() -> Res {
 fn transpose_shrd_mem() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/transpose_shrd_mem.desc")?
+        descend::compile(
+            "examples/infer/transpose_shrd_mem.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -24,7 +27,7 @@ fn transpose_shrd_mem() -> Res {
 fn matmul() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/matmul.desc")?
+        descend::compile("examples/infer/matmul.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -32,7 +35,7 @@ fn matmul() -> Res {
 fn scale_vec() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/scale_vec.desc")?
+        descend::compile("examples/infer/scale_vec.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -40,7 +43,7 @@ fn scale_vec() -> Res {
 fn reverse_vec() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/reverse_vec.desc")?
+        descend::compile("examples/infer/reverse_vec.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -49,7 +52,10 @@ fn reverse_vec() -> Res {
 fn bitonic_sort() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/bitonic_sort/bitonic_sort.desc")?
+        descend::compile(
+            "examples/infer/bitonic_sort/bitonic_sort.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -62,7 +68,7 @@ fn scan() -> Res {
     );
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/scan.desc")?
+        descend::compile("examples/infer/scan.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -70,7 +76,10 @@ fn scan() -> Res {
 fn reduce_shared_mem() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/reduce_shared_mem.desc")?
+        descend::compile(
+            "examples/infer/reduce_shared_mem.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -78,7 +87,10 @@ fn reduce_shared_mem() -> Res {
 fn vlc_encode() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/huffman/vlc_encode.desc")?
+        descend::compile(
+            "examples/infer/huffman/vlc_encode.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -86,7 +98,10 @@ fn vlc_encode() -> Res {
 fn vlc_encode_cg() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/huffman/vlc_encode_cg.desc")?
+        descend::compile(
+            "examples/infer/huffman/vlc_encode_cg.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -94,7 +109,10 @@ fn vlc_encode_cg() -> Res {
 fn vlc_encode_reuse() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/huffman/vlc_encode_reuse.desc")?
+        descend::compile(
+            "examples/infer/huffman/vlc_encode_reuse.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -102,7 +120,10 @@ fn vlc_encode_reuse() -> Res {
 fn histogram() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/huffman/histogram.desc")?
+        descend::compile(
+            "examples/infer/huffman/histogram.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -110,7 +131,7 @@ fn histogram() -> Res {
 fn tree_reduce() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/tree_reduce.desc")?
+        descend::compile("examples/infer/tree_reduce.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -118,21 +139,24 @@ fn tree_reduce() -> Res {
 fn vector_add() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/vec_add.desc")?
+        descend::compile("examples/infer/vec_add.desc", descend::Backend::Cuda)?
     ))
 }
 
 #[ignore]
 #[test]
 fn bfs() -> Res {
-    Ok(println!("{}", descend::compile("examples/infer/bfs.desc")?))
+    Ok(println!(
+        "{}",
+        descend::compile("examples/infer/bfs.desc", descend::Backend::Cuda)?
+    ))
 }
 
 #[test]
 fn sgemm() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/sgemm.desc")?
+        descend::compile("examples/infer/sgemm.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -140,7 +164,10 @@ fn sgemm() -> Res {
 fn shrd_mem_acc_equiv_exec() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/shrd_mem_acc_equiv_exec.desc")?
+        descend::compile(
+            "examples/shrd_mem_acc_equiv_exec.desc",
+            descend::Backend::Cuda
+        )?
     ))
 }
 
@@ -148,7 +175,7 @@ fn shrd_mem_acc_equiv_exec() -> Res {
 fn sssp_ffi_unsafe() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/sssp-ffi.desc")?
+        descend::compile("examples/infer/sssp-ffi.desc", descend::Backend::Cuda)?
     ))
 }
 
@@ -156,6 +183,6 @@ fn sssp_ffi_unsafe() -> Res {
 fn jacobisvd() -> Res {
     Ok(println!(
         "{}",
-        descend::compile("examples/infer/jacobisvd.desc")?
+        descend::compile("examples/infer/jacobisvd.desc", descend::Backend::Cuda)?
     ))
 }
